DCS-86-3 Refactor service folders
Move reporting stuff into a single folder.

Rename reports folder to draft
