+++ server/routes/addInvolvedStaff.ts
@@ -3,7 +3,7 @@ import { Request, Response } from 'express'
 import { properCaseFullName } from '../utils/utils'
 import { nextPaths, paths } from '../config/incident'
 import { SystemToken } from '../types/uof'
-import DraftReportService, { AddStaffResult } from '../services/report/draftReportService'
+import DraftReportService, { AddStaffResult } from '../services/drafts/draftReportService'
 
 const SubmitType = {
   SAVE_AND_CONTINUE: 'save-and-continue',
