+++ server/services/reporting/heatmapBuilder.test.ts
