+++ server/services/reporting/incidentCountAggregator/aggregatorFunctions.test.ts
@@ -1,5 +1,5 @@
 import { invertGroupings, aggregatorFactory } from './aggregatorFunctions'
-import { PrisonerDetail } from '../../data/elite2ClientBuilderTypes'
+import { PrisonerDetail } from '../../../data/elite2ClientBuilderTypes'
 
 describe('Aggregator Functions', () => {
   describe('invertGroupings', () => {
