+++ server/services/reporting/incidentCountAggregator/index.ts
